x86: normalize types used with _mm{,256}_set?_epi64x
w/clang -fsanitize=integer fixes warnings of the form:
implicit conversion from type 'uint64_t' (aka 'unsigned long') of value
15120695205995520509 (64-bit, unsigned) to type 'long long' changed the
value to -3326048867714031107 (64-bit, signed)

Bug: aomedia:3136
Bug: b/229626362

+++ aom_dsp/simd/v128_intrinsics_x86.h
@@ -583,8 +583,8 @@ SIMD_INLINE v128 v128_shr_u64(v128 a, unsigned int c) {
 
 SIMD_INLINE v128 v128_shr_s64(v128 a, unsigned int c) {
   // _mm_sra_epi64 is missing in gcc?
-  return v128_from_64((int64_t)v64_u64(v128_high_v64(a)) >> c,
-                      (int64_t)v64_u64(v128_low_v64(a)) >> c);
+  return v128_from_64((uint64_t)((int64_t)v64_u64(v128_high_v64(a)) >> c),
+                      (uint64_t)((int64_t)v64_u64(v128_low_v64(a)) >> c));
   // return _mm_sra_epi64(a, _mm_cvtsi32_si128((int)c));
 }
